registration overhaul (WIP)

+++ app/controllers/after_signup_controller.rb
@@ -16,20 +16,39 @@ class AfterSignupController < ApplicationController
   end
 
   def create
-    # create user
-    user = User.create email: params[:temporary_user][:email], password: params[:temporary_user][:password], display_name: params[:temporary_user][:display_name]
-    user.socialproviders.create provider: session[:user_provider], uid: session[:user_uid], display_name: params[:temporary_user][:display_name], email: params[:temporary_user][:email]
+    # create new user and validate data
+    @user = User.new(email: params[:temporary_user][:email], password: params[:temporary_user][:password], display_name: params[:temporary_user][:display_name])
+    @user.validate_display_name(params[:temporary_user][:display_name])
+    @user.validate_email(params[:temporary_user][:email], params[:temporary_user][:email_confirm])
+    @user.validate_password(params[:temporary_user][:password], params[:temporary_user][:password_confirm])
 
-    # create flash message
-    flash[:notice] = t('authentication.social.created', provider: session[:user_provider].capitalize)
+    if @user.errors.empty?
+      # save user and link social provider with account
+      @user.save
+      @user.socialproviders.create provider: session[:user_provider], uid: session[:user_uid], display_name: params[:temporary_user][:display_name], email: params[:temporary_user][:email]
 
-    # delete session variables
-    session.delete(:user_display_name)
-    session.delete(:user_email)
-    session.delete(:user_provider)
-    session.delete(:user_uid)
+      # create flash message
+      flash[:notice] = t('authentication.social.created', provider: session[:user_provider].capitalize)
 
-    # sign in and redirect
-    sign_in_and_redirect :user, user
+      # delete session variables
+      session.delete(:user_display_name)
+      session.delete(:user_email)
+      session.delete(:user_provider)
+      session.delete(:user_uid)
+
+      # sign in and redirect
+      sign_in_and_redirect :user, @user
+    else
+      # add error messages
+      flash[:error] = ''
+      flash[:error] += @user.errors[:display_name].first.to_s if @user.errors[:display_name]
+      flash[:error] += @user.errors[:email].first.to_s if @user.errors[:email]
+      flash[:error] += @user.errors[:email_confirm].first.to_s if @user.errors[:email_confirm]
+      flash[:error] += @user.errors[:password].first.to_s if @user.errors[:password]
+      flash[:error] += @user.errors[:password_confirm].first.to_s if @user.errors[:password_confirm]
+
+      # redirect to after signup path
+      redirect_to after_signup_index_path
+    end
   end
 end
